requirements.txt
Flask
Pillow
numpy
easyocr
